Move CachedResponse class into its own file
